feat: 버튼 공통화
