+++ app/src/main/java/com/dogandpigs/fitnote/presentation/login/LoginScreen.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import androidx.navigation.compose.rememberNavController
 import com.dogandpigs.fitnote.presentation.base.FigmaPreview
+import com.dogandpigs.fitnote.presentation.ui.component.CompleteButton
 import com.dogandpigs.fitnote.presentation.ui.component.FitNoteScaffold
 import com.dogandpigs.fitnote.presentation.ui.theme.FitNoteTheme
 
@@ -105,23 +106,8 @@ private fun Login(
                     }
                 }
             }
-
-            OutlinedButton(
-                modifier = Modifier
-                    .width(100.dp)
-                    .padding(0.dp, 10.dp)
-                    .background(Color.Transparent)
-                    .align(Alignment.BottomCenter),
-                onClick = {
-//                        viewModel.login(email.text, pwd.text)
-                    navigateToHome()
-                },
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color.White, contentColor = Color.Black
-                )
-            ) {
-                Text(text = "다음")
-            }
+    
+            CompleteButton("저장", onClick = {})
         }
     }
 }
